% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresionlinealmultiple.R
\name{multi_reg}
\alias{multi_reg}
\title{Title: Función de regresión lineal múltiple}
\usage{
multi_reg(data, dependent_var, independent_vars)
}
\arguments{
\item{data}{Data to be used to fit the multiple linear regression model.}

\item{dependent_var}{La variable dependiente que se desea predecir mediante el modelo.}

\item{independent_vars}{Vector with the independent variables that will be used to fit the multiple linear regression model.}
}
\value{
Summary of the fitted model and a plot of fitted values
}
\description{
Title: Función de regresión lineal múltiple
}
\examples{
df <- data.frame(x1 = c(1,2,3,4,5),
                x2 = c(3,4,5,6,7),
                y = c(6,9,12,15,18))
multi_reg(df, "y", c("x1", "x2"))
}
