% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlacion.R
\name{my_corr_plot}
\alias{my_corr_plot}
\title{Title: Funci贸n de correlaci贸n}
\usage{
my_corr_plot(data, x_var, y_var)
}
\arguments{
\item{data}{Data frame that has the variables x_var and y_var.}

\item{x_var}{The name of the data frame data column to be used as the independent variable in the scatter plot.}

\item{y_var}{the name of the data frame data column to use as the dependent variable in the scatter plot.}
}
\value{
The value of the calculated correlation
}
\description{
Title: Funci贸n de correlaci贸n
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(2, 4, 6, 8, 10)
df <- data.frame(x, y)
my_corr_plot(df, "x", "y")
}
