% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regresionlineals.R
\name{reg_lineal}
\alias{reg_lineal}
\title{Title: Funci贸n de regresi贸n lineal simple}
\usage{
reg_lineal(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
Simple linear regression fit
}
\description{
Title: Funci贸n de regresi贸n lineal simple
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(3, 5, 7, 9, 11)
reg_lineal(x, y)

}
